deepsight_backend/deep_researcher_agent/requirements.txt
dspy_ai==2.4.9
wikipedia==1.4.0
sentence-transformers
toml
langchain-text-splitters
trafilatura
langchain-huggingface
qdrant-client
langchain-qdrant
numpy==1.26.4
litellm==1.59.3
diskcache
rank_bm25==0.2.2
tabulate
tavily-python
google-generativeai
# Dependencies for image extractor
yt-dlp
ffmpeg-python
pillow
imagehash
open_clip_torch
openai
easyocr

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pandas>=1.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.25.0